Move worker launch scheduling out of the router
