+++ src/kitaru/server/adapters/rest/routers/imports.py
@@ -13,15 +13,12 @@
 #  permissions and limitations under the License.
 """Import routes."""
 
-import logging
 from typing import Annotated
 
 from fastapi import APIRouter, BackgroundTasks, Depends, status
-from pydantic import SecretStr
 
 from kitaru.api_models.v1.imports import ImportCreateRequest
 from kitaru.api_models.v1.job import JobResponse
-from kitaru.api_models.v1.worker import WorkerRuntime
 from kitaru.server.adapters.auth.auth_service import AuthService
 from kitaru.server.adapters.rest.dependencies import (
     authorize,
@@ -35,15 +32,12 @@ from kitaru.server.adapters.rest.mapping.imports import import_create_to_command
 from kitaru.server.adapters.rest.mapping.jobs import job_to_response
 from kitaru.server.adapters.rest.responses import error_responses
 from kitaru.server.adapters.rest.route import KitaruAPIRoute, idempotent
+from kitaru.server.adapters.rest.worker_launch import schedule_worker_launch
 from kitaru.server.api.config import APISettings
 from kitaru.server.application.interfaces.worker_launcher import WorkerLauncher
 from kitaru.server.application.models.auth import AuthContext
-from kitaru.server.application.models.worker import WorkerLaunch
 from kitaru.server.application.services.job_service import JobService
 from kitaru.server.application.services.worker_service import WorkerService
-from kitaru.server.domain.task import Task
-
-logger = logging.getLogger(__name__)
 
 router = APIRouter(route_class=KitaruAPIRoute)
 
@@ -85,7 +79,7 @@ async def create_import(
     command = import_create_to_command(body)
     job, task = await service.create_import(command, actor=actor)
     if launcher is not None:
-        await _schedule_worker_launch(
+        await schedule_worker_launch(
             task,
             worker_service,
             auth_service,
@@ -95,81 +89,3 @@ async def create_import(
             actor,
         )
     return job_to_response(job)
-
-
-async def _schedule_worker_launch(
-    task: Task,
-    worker_service: WorkerService,
-    auth_service: AuthService,
-    launcher: WorkerLauncher,
-    settings: APISettings,
-    background_tasks: BackgroundTasks,
-    actor: AuthContext,
-) -> None:
-    """Register a worker for the task and launch it after the response.
-
-    A live worker whose scope covers the task suppresses the launch.
-
-    Args:
-        task: Task the worker is started for.
-        worker_service: Worker service.
-        auth_service: Authentication service for the current request.
-        launcher: Worker launcher.
-        settings: API settings for this process.
-        background_tasks: Tasks run after the response is sent.
-        actor: Caller context.
-    """
-    if await worker_service.is_covered(task):
-        return
-    worker = await worker_service.register_ephemeral_worker(
-        task,
-        WorkerRuntime(platform=settings.WORKER_LAUNCHER.backend.value),
-        actor=actor,
-    )
-    issued = auth_service.issue_worker_token(
-        worker_id=worker.id,
-        account_id=actor.account.id,
-        lifetime_seconds=_get_worker_token_lifetime_seconds(settings),
-    )
-    launch = WorkerLaunch(
-        worker_id=worker.id,
-        worker_token=SecretStr(issued.token),
-        server_url=settings.SERVER_URL,
-        job_id=task.job_id,
-    )
-    # The route class commits the session before the response goes out and
-    # Starlette runs background tasks after it, so the worker row is
-    # committed before the sandbox starts.
-    background_tasks.add_task(_launch_worker, launcher, launch)
-
-
-def _get_worker_token_lifetime_seconds(settings: APISettings) -> int:
-    """Compute the lifetime of an ephemeral worker token.
-
-    Args:
-        settings: API settings for this process.
-
-    Returns:
-        Sandbox timeout plus the task token expiry leeway, in seconds.
-    """
-    # Modal is the only launcher backend, and settings validation requires
-    # its sub-model to be set under that backend.
-    modal = settings.WORKER_LAUNCHER.modal
-    assert modal is not None
-    return modal.timeout_seconds + settings.TASK_TOKEN_EXPIRY_LEEWAY_SECONDS
-
-
-async def _launch_worker(launcher: WorkerLauncher, launch: WorkerLaunch) -> None:
-    """Launch a worker, logging a failure instead of raising.
-
-    Args:
-        launcher: Worker launcher.
-        launch: Worker launch.
-    """
-    try:
-        await launcher.launch(launch)
-    except Exception:
-        # Nothing retries here, a lost launch is left to the liveness window.
-        logger.exception(
-            "Failed to launch worker %s for job %s.", launch.worker_id, launch.job_id
-        )
